+++ packages/logger/src/levels.ts
@@ -1,0 +1,10 @@
+import type { LogLevel } from "./types";
+
+export const LOG_LEVELS: Record<LogLevel, number> = {
+  trace: 10,
+  debug: 20,
+  info: 30,
+  warn: 40,
+  error: 50,
+  fatal: 60,
+};
