+++ packages/logger/src/util/time.ts
@@ -1,0 +1,3 @@
+export function nowIso(): string {
+  return new Date().toISOString();
+}
